+++ packages/nuxt/playground/components/AppSidebar.vue
@@ -86,6 +86,10 @@ const data = {
           url: '/components/collapsible',
         },
         {
+          title: 'Combobox',
+          url: '/components/combobox',
+        },
+        {
           title: 'Dialog',
           url: '/components/dialog',
         },