+++ packages/nuxt/src/runtime/components/combobox/ComboboxEmpty.vue
@@ -1,0 +1,28 @@
+<script setup lang="ts">
+import type { NComboboxEmptyProps } from '../../types'
+import { ComboboxEmpty } from 'reka-ui'
+import { computed } from 'vue'
+import { cn } from '../../utils'
+
+const props = defineProps<NComboboxEmptyProps>()
+
+const delegatedProps = computed(() => {
+  const { class: _, ...delegated } = props
+
+  return delegated
+})
+</script>
+
+<template>
+  <ComboboxEmpty
+    data-slot="combobox-empty"
+    v-bind="delegatedProps"
+    :class="cn(
+      'combobox-empty',
+      props.una?.comboboxEmpty,
+      props.class,
+    )"
+  >
+    <slot />
+  </ComboboxEmpty>
+</template>